Refactor: Move Dependency into entities
